rename: image와 style 파일 src안으로 변경
