+++ docs/REQUIREMENTS.md
@@ -75,3 +75,6 @@
 - [ ] 전역 객체인 Store의 값을 함부로 set 할 수 있다는 점이 위험한데, 이 부분을 잘 생각해본다.
 - [ ] 전역에서 동작하는 UI 컨트롤러의 구현을 각각의 담당 역할로 나누어 구현한다. (Store에 너무 몰려
       있다.)
+- [ ] 디바운스 / 스로틀링 사용해서 요청 최적화 하기
+- [ ] skeleton UI 최적화 하기
+- [ ]
