+++ src/index.js
@@ -1,6 +1,6 @@
 // import Movie from './domain/Movie';
-import '../templates/reset.css';
-import '../templates/common.css';
+import './styles/reset.css';
+import './styles/common.css';
 
 import App from './App';
