+++ src/styles/common.css
@@ -171,6 +171,6 @@ header .search-box > .search-button {
   width: 14px;
   border: 0;
   text-indent: -1000rem;
-  background: url('./search_button.png') transparent no-repeat 0 1px;
+  background: url('../images/search_button.png') transparent no-repeat 0 1px;
   background-size: contain;
 }
